Extract *-spring-boot-starter modules doing autoconfiguration
