+++ core-starter/build.gradle
@@ -1,0 +1,11 @@
+ext.projectName = "TwTasks Core - Spring Boot Starter"
+ext.projectDescription = "Autoconfigures core for spring boot application"
+ext.projectArtifactName = "tw-tasks-core-starter"
+
+apply from: '../build.common.gradle'
+apply from: '../build.core-dependencies.gradle'
+apply from: '../build.publish.gradle'
+
+dependencies {
+    api project(":core")
+}
