+++ core-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.transferwise.tasks.core.autoconfigure.TwTasksCoreAutoConfiguration
+org.springframework.context.ApplicationListener=com.transferwise.tasks.TwTasksApplicationListener
