--- ext-incidents/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.transferwise.tasks.ext.incidents.config.TwTasksExtIncidentsAutoConfiguration
