+++ ext-kafka-publisher-test/build.gradle
@@ -1,4 +1,4 @@
-ext.projectName = "TwTasks Extension - Kafka Publisher (Test Components)"
+ext.projectName = "TwTasks Extension - Kafka Publisher - Test Components"
 ext.projectDescription = "Set of components that simply testing around kafka messages publishing"
 ext.projectArtifactName = "tw-tasks-ext-kafka-publisher-test"
