+++ ext-management-starter/build.gradle
@@ -1,0 +1,9 @@
+ext.projectName = "TwTasks Extension - Management - Spring Boot Starter"
+ext.projectDescription = "Autoconfigures extension for spring boot application"
+ext.projectArtifactName = "tw-tasks-ext-management-starter"
+
+apply from: '../build.ext.gradle'
+
+dependencies {
+    api project(":ext-management")
+}
