+++ ext-management-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.transferwise.tasks.ext.management.autoconfigure.TwTasksExtManagementAutoConfiguration
