--- ext-management/src/main/java/com/transferwise/tasks/ext/management/config/TwTasksExtManagementAutoConfiguration.java
@@ -1,22 +1,0 @@
-package com.transferwise.tasks.ext.management.config;
-
-import com.transferwise.tasks.management.TasksManagementPortController;
-import com.transferwise.tasks.management.TasksManagementService;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class TwTasksExtManagementAutoConfiguration {
-
-  @Bean
-  public TasksManagementService twTasksTasksManagementService() {
-    return new TasksManagementService();
-  }
-
-  @Bean
-  @ConditionalOnMissingBean(TasksManagementPortController.class)
-  public TasksManagementPortController twTasksTasksManagementPortController() {
-    return new TasksManagementPortController();
-  }
-}
