+++ settings.gradle
@@ -10,11 +10,16 @@ pluginManagement {
 rootProject.name = 'tw-tasks-executor'
 
 include 'core'
+include 'core-starter'
 include 'core-test'
 include 'ext-incidents'
+include 'ext-incidents-starter'
 include 'ext-kafka-listener'
+include 'ext-kafka-listener-starter'
 include 'ext-kafka-publisher'
+include 'ext-kafka-publisher-starter'
 include 'ext-kafka-publisher-test'
 include 'ext-management'
+include 'ext-management-starter'
 include 'integration-tests'
 include 'demoapp'
